requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
googlemaps==4.10.0
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.0